requirements.txt
fastapi==0.104.1
pydantic>=1.10.0,<2.0.0
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
vosk==0.3.45
moviepy==1.0.3
redis==4.6.0
aiofiles==23.2.1
python-dotenv==1.0.0
pydub==0.25.1
sqlalchemy==1.4.42
psycopg2-binary==2.9.9
alembic==1.12.0
databases[postgresql]==0.8.0
asyncpg==0.29.0
